+++ crates/router/src/core/payments/operations.rs
@@ -11,6 +11,7 @@ pub mod payment_start;
 pub mod payment_status;
 pub mod payment_update;
 pub mod payments_incremental_authorization;
+pub mod tax_calculation;
 
 use api_models::enums::FrmSuggestion;
 use async_trait::async_trait;
@@ -24,6 +25,7 @@ pub use self::{
     payment_response::PaymentResponse, payment_session::PaymentSession,
     payment_start::PaymentStart, payment_status::PaymentStatus, payment_update::PaymentUpdate,
     payments_incremental_authorization::PaymentIncrementalAuthorization,
+    tax_calculation::PaymentSessionUpdate,
 };
 use super::{helpers, CustomerDetails, OperationSessionGetters, OperationSessionSetters};
 use crate::{
@@ -177,6 +179,20 @@ pub trait Domain<F: Clone, R, D>: Send + Sync {
         _merchant_account: &domain::BusinessProfile,
         _key_store: &domain::MerchantKeyStore,
         _mandate_type: Option<api_models::payments::MandateTransactionType>,
+    ) -> CustomResult<(), errors::ApiErrorResponse> {
+        Ok(())
+    }
+
+    #[allow(clippy::too_many_arguments)]
+    async fn payments_dynamic_tax_calculation<'a>(
+        &'a self,
+        _state: &SessionState,
+        _payment_data: &mut D,
+        _should_continue_confirm_transaction: &mut bool,
+        _connector_call_type: &ConnectorCallType,
+        _business_profile: &domain::BusinessProfile,
+        _key_store: &domain::MerchantKeyStore,
+        _merchant_account: &domain::MerchantAccount,
     ) -> CustomResult<(), errors::ApiErrorResponse> {
         Ok(())
     }